add deterministic generation of state constant
So far, state values constants were generated randomly. If new state is added, constants are changing which is not desirable if some modules depends on the specific value to check (e.g., is state as expected)?

Now, the constant is generated deterministically from the state name and is independent from other states